Added code-analyzer, started with verification and fixed code to fit analzer (mostly - 95%)

+++ SkyBot/AccessLevel.cs
@@ -4,7 +4,7 @@ using System.Text;
 
 namespace SkyBot
 {
-    public enum AccessLevel : short
+    public enum AccessLevel : int
     {
         User,
         VIP,
